// src/types/api/daily-transaction.ts
export type DailyTransactionItem = {
  ReservationId: string
  CustomerId?: string | null
  BusinessId?: string | null
  ShowId?: string | null
  ShowDetId?: string | null
  Dinner?: string | null
  CreatedBy?: string | null
  CreateDt?: string | null
  Total?: number | null
  ResSource?: string | null
  ResStatus?: string | null
  PartyNo?: number | null
  CheckedIn?: number | null
  Price?: number | null
  Promo?: string | null
  NumPasses?: number | null
  PendingStatus?: string | null
  Notes?: string | null
  FirstName?: string | null
  LastName?: string | null
  BusinessName?: string | null
  AreaCode?: string | null
  Phone1?: string | null
  Phone2?: string | null
  Source?: string | null
  Section?: string | null
  PymtLName?: string | null
  PymtFName?: string | null
  FirstName1?: string | null
  LastName1?: string | null
  CCType?: string | null
  PymtStatus?: string | null
  PymtType?: string | null
  Amount?: number | null
}